move network transport interface to network file
